pypi.bbclass: Allow fetching without requiring the hash
This allow the fetching of packages in a backward compatible way[1],
using un-hashed packages.

1. https://bitbucket.org/pypa/pypi/issues/438/backwards-compatible-un-hashed-package

A equivalent fix is available on master in commit 7edad4d (pypi:
update the base pypi url to use the package info without requiring the
hash) so this is really a backport but without unrelated changes.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
Acked-by: Tim Orling (Intel) <ticotimo@gmail.com>
Tested-by: Tim Orling (Intel) <ticotimo@gmail.com>
Acked-by: Armin Kuster <akuster808@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>